Issue 1: Fixed; wiki page for 'malus' not showing up in candidate_pages
Fixed the issue regarding ‘a few words from Wiktionary not getting into candidate_pages folder’s files’. A hardcoded string in the file expected only <h3> header files to have Noun, Adjective etc POS but in case of ‘malus’, the POS was in <h4> header. Added an ‘or’ statement to include <h4> header as well.